requirements.txt
amqp==5.1.0
arabic-reshaper==2.1.3
asgiref==3.5.2
async-timeout==4.0.2
Babel==2.10.3
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.2.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
dj-crontab==0.8.0
Django==4.1.3
django-appconf==1.0.5
django-autoslug==1.9.8
django-ckeditor==6.4.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.8.1
django-extensions==2.2.5
django-filter==2.4.0
django-js-asset==2.0.0
django-meta==2.0.0
django-modeltranslation==0.18.2
django-money==3.0.0
django-oauth-toolkit==2.2.0
django-phonenumber-field==7.0.1
django-push-notifications==3.0.0
django-request==1.5.6
django-rest-auth==0.9.5
django-rosetta==0.9.8
django-social-share==2.2.1
django-versatileimagefield==2.2
django-zipfile==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==19.9.0
html5lib==1.1
idna==3.4
jwcrypto==1.4.2
kombu==5.2.4
# mysqlclient==2.0.3
newrelic==7.2.4.171
oauth2-provider==0.0
oauthlib==3.2.2
packaging==21.3
pdf2image==1.16.0
phonenumbers==8.13.1
Pillow==9.3.0
platformdirs==2.2.0
polib==1.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.2
py-moneyed==2.0
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.0
# python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2022.7
qrcode==7.3
redis==4.2.2
reportlab==3.6.1
requests==2.28.1
requests-oauthlib==1.3.0
sentry-sdk==1.5.7
six==1.16.0
social==0.1.dev3
social-auth-app-django==3.1.0
social-auth-core==4.1.0
sqlparse==0.4.3
stripe==2.61.0
typing_extensions==4.3.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.14.1
xhtml2pdf==0.2.5
XlsxWriter==3.0.2
zope.event==4.5.0
zope.interface==5.4.0
